+++ packages/core/package.json
@@ -59,6 +59,7 @@
     "@anthropic-ai/sdk": "0.32.1",
     "@aws-sdk/client-s3": "3.850.0",
     "@aws-sdk/s3-request-presigner": "3.850.0",
+    "@google-cloud/storage": "^7.7.0",
     "@codesandbox/sdk": "0.6.2",
     "@kubernetes/client-node": "1.0.0",
     "@latitude-data/compiler": "workspace:^",
